+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+v0.1.2
+======
+* Fixed issue with Streamer#send not handling null bytes well for binary data;
+  required adjusting Librtmp::FFI#RTMP_Write() implementation to expect a
+  pointer instead of a string.
+* Stubbed out implementation for converting a Hash sent to Streamer#send into
+  an AMF Object to send over the stream.
+
 v0.1.1
 ======